[VSF2-213] Generators moved to separate dirs
