+++ src/DataGenerator/Generator/Bulk/Collection/WishlistProductCollectionBulkGeneratorInterface.php
@@ -8,8 +8,11 @@
 
 declare(strict_types=1);
 
-namespace BitBag\SyliusVueStorefront2Plugin\DataGenerator\Generator\Bulk;
+namespace BitBag\SyliusVueStorefront2Plugin\DataGenerator\Generator\Bulk\Collection;
 
-interface ProductBulkGeneratorInterface extends BulkGeneratorInterface
+use BitBag\SyliusVueStorefront2Plugin\DataGenerator\Generator\Bulk\BulkGeneratorInterface;
+
+interface WishlistProductCollectionBulkGeneratorInterface extends BulkGeneratorInterface
 {
+    const LIMIT = 100;
 }
