+++ src/DataGenerator/Generator/Bulk/Entity/ProductBulkGeneratorInterface.php
@@ -8,9 +8,10 @@
 
 declare(strict_types=1);
 
-namespace BitBag\SyliusVueStorefront2Plugin\DataGenerator\Generator\Bulk;
+namespace BitBag\SyliusVueStorefront2Plugin\DataGenerator\Generator\Bulk\Entity;
 
-interface ProductTaxonCollectionBulkGeneratorInterface extends BulkGeneratorInterface
+use BitBag\SyliusVueStorefront2Plugin\DataGenerator\Generator\Bulk\BulkGeneratorInterface;
+
+interface ProductBulkGeneratorInterface extends BulkGeneratorInterface
 {
-    const LIMIT = 100;
 }
