+++ src/DataGenerator/Generator/Bulk/Entity/WishlistBulkGenerator.php
@@ -8,7 +8,7 @@
 
 declare(strict_types=1);
 
-namespace BitBag\SyliusVueStorefront2Plugin\DataGenerator\Generator\Bulk;
+namespace BitBag\SyliusVueStorefront2Plugin\DataGenerator\Generator\Bulk\Entity;
 
 final class WishlistBulkGenerator extends AbstractBulkGenerator implements WishlistBulkGeneratorInterface
 {
